+++ kurtosis-module/impl/prelaunch_data_generator/prelaunch_data_generator_launcher.go
@@ -1,11 +1,13 @@
 package prelaunch_data_generator
 
 import (
+	"fmt"
 	"github.com/kurtosis-tech/eth2-merge-kurtosis-module/kurtosis-module/impl/prelaunch_data_generator/cl_genesis"
 	"github.com/kurtosis-tech/eth2-merge-kurtosis-module/kurtosis-module/impl/prelaunch_data_generator/cl_validator_keystores"
 	"github.com/kurtosis-tech/kurtosis-core-api-lib/api/golang/lib/enclaves"
 	"github.com/kurtosis-tech/kurtosis-core-api-lib/api/golang/lib/services"
 	"github.com/kurtosis-tech/stacktrace"
+	"time"
 )
 
 const (
@@ -14,7 +16,7 @@ const (
 	// It's only a Kurtosis image because the original repo doesn't publish Docker images
 	image = "skylenet/ethereum-genesis-generator:latest"
 
-	serviceId services.ServiceID = "prelaunch-data-generator"
+	serviceIdPrefix = "prelaunch-data-generator-"
 )
 
 // We use Docker exec commands to run the commands we need, so we override the default
@@ -32,19 +34,35 @@ type PrelaunchData struct {
 
 func LaunchPrelaunchDataGenerator(
 	enclaveCtx *enclaves.EnclaveContext,
+	filesArtifactMountpoints map[services.FilesArtifactID]string,
+
+	/*
 	networkId string,
 	depositContractAddress string,
 	totalTerminalDifficulty uint64,
 	preregisteredValidatorKeysMnemonic string,
+
+	 */
 ) (
-	*PrelaunchDataGeneratorContext,
+	*services.ServiceContext,
 	error,
 ) {
-	serviceCtx, err := enclaveCtx.AddService(serviceId, getContainerConfig)
+	containerConfigSupplier := getContainerConfigSupplier(filesArtifactMountpoints)
+
+	serviceId := services.ServiceID(fmt.Sprintf(
+		"%v%v",
+		serviceIdPrefix,
+		time.Now().Unix(),
+	))
+
+	serviceCtx, err := enclaveCtx.AddService(serviceId, containerConfigSupplier)
 	if err != nil {
-		return nil, stacktrace.Propagate(err, "An error occurred launching the prelaunch data generator container with service ID '%v'", serviceId)
+		return nil, stacktrace.Propagate(err, "An error occurred launching the prelaunch data generator container with service ID '%v'", serviceIdPrefix)
 	}
 
+	return serviceCtx, nil
+
+	/*
 	result := newPrelaunchDataGeneratorContext(
 		serviceCtx,
 		networkId,
@@ -53,14 +71,22 @@ func LaunchPrelaunchDataGenerator(
 		preregisteredValidatorKeysMnemonic,
 	)
 	return result, nil
+
+	 */
 }
 
-func getContainerConfig(privateIpAddr string, sharedDir *services.SharedPath) (*services.ContainerConfig, error) {
-	containerConfig := services.NewContainerConfigBuilder(
-		image,
-	).WithEntrypointOverride(
-		entrypointArgs,
-	).Build()
+func getContainerConfigSupplier(
+	filesArtifactMountpoints map[services.FilesArtifactID]string,
+) func(privateIpAddr string) (*services.ContainerConfig, error) {
+	return func(privateIpAddr string) (*services.ContainerConfig, error) {
+		containerConfig := services.NewContainerConfigBuilder(
+			image,
+		).WithEntrypointOverride(
+			entrypointArgs,
+		).WithFiles(
+			filesArtifactMountpoints,
+		).Build()
 
-	return containerConfig, nil
+		return containerConfig, nil
+	}
 }